Fixed type-casting 32<->64 bit following the >2G commit
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@2034 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ modules/svg_in/svg_in.c
@@ -46,7 +46,7 @@ typedef struct
 
 static Bool svg_check_download(SVGIn *svgin)
 {
-	u32 size;
+	u64 size;
 	FILE *f = gf_f64_open(svgin->file_name, "rb");
 	gf_f64_seek(f, 0, SEEK_END);
 	size = gf_f64_tell(f);
